requirements.txt
Django==2.2.7
django-chartjs==1.5.0
factory-boy==2.12.0
Faker==2.0.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
